Main

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

data ModoJogo

Constructors

MenuNovoJogo
Opção jogar selecionada no menu

MenuContinuarJogo
Opção 'continuar a jogar'

MenuControlos
Opção controlos

MenuCreditos
Opção créditos

MenuSair
Opção sair selecionada no menu

Jogar
Enquanto joga

Controlos
Janela dos controlos

Creditos
Parte dos créditos

Perdeu MenuPerdeu
Menu quando o jogo termina

Pausa MenuPausa
Menu de pausa

data MenuPausa

Várias opções do menu de pausa.

Constructors

VoltarJogo

GuardarSair

SairPausa

data MenuPerdeu

Várias opções do menu de fim de jogo.

Constructors

JogarDeNovo

PerdeuSair

type Score = Int

Pontuação do jogador.

type Tempo = Int

Contador auxiliar de tempo.

type Estado = (ModoJogo, Jogo, Score, Score, [Picture], Direcao, Tempo, Score)

Estado do jogo para guardar informações: o modo de jogo, o jogo, a pontuação e a melhor pontuação do jogo, a lista de imagens usadas, a direção do último movimento, o contador do tempo e a melhor pontuação que o jogador já teve.

gerarMapa :: [Int] -> Mapa -> Int -> Mapa

Função para gerar um mapa aleatório.

mainDisplay :: Display

Janela do jogo.

baseMapa :: Mapa

Base do mapa fixa, para melhorar a jogabilidade.

estadoInicial :: Mapa -> [Picture] -> Estado

Estado inicial do jogo.

drawState :: Estado -> IO Picture

Função para desenhar o jogo.

drawLines :: (Int, Int) -> [(Terreno, [Obstaculo])] -> [Picture] -> [Picture]

Função para desenhar cada linha linha do mapa.

drawObstacles :: Int -> (Terreno, [Obstaculo]) -> [Picture] -> [Picture]

Função para desenhar os obstáculos numa linha do mapa.

drawMiddle :: Int -> Int -> [[Obstaculo]] -> [Picture] -> [Picture]

Função para desenhar os obstáculos da Estrada excluíndo, se for o caso, os obstáculos das pontas.

event :: Event -> Estado -> IO Estado

Função para reagir às teclas pressionadas pelo utilizador.

time :: Float -> Estado -> IO Estado

Função que define a passagem do tempo no jogo e que executa algo à medida que o tempo passa.

main :: IO ()

Função para correr o jogo.