Tarefa1_2022li1g067

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

Função principal que corre tudo

terrenoValido :: Terreno -> Bool

Valida um terreno, não pode ter velocidade 0

obstaculoValido :: Largura -> Terreno -> [Obstaculo] -> Bool

Ponto 1 e 5, verifica se tem por onde avançar e se os obstáculos estão nos terrenos corretos

lIgualv :: Largura -> [Obstaculo] -> Bool

Ponto 6, verifica se a largura é igual ao número de obstáculos

obsV2 :: Terreno -> [Obstaculo] -> Bool

Ponto 3 e 4, verifica se não tem obstáculos consecutivos a mais

ols :: Terreno -> [[Obstaculo]] -> [Int]

Auxiliar da função obsV2, conta quantos carros e troncos seguidos existe

terrenoSeguidos :: [Terreno] -> Bool

Ponto 7, não permite mais de um número seguido de rio, estrada e relva

tRS :: [(Terreno, [Obstaculo])] -> [Terreno]

Usada na função principal para converter [(Terreno, [Obstaculo])] em [Terreno]

tAux :: [Terreno] -> [Terreno]

Atribui o valor 0 às velocidades para depois ser possivel agrupar

tAux2 :: [[Terreno]] -> Bool

Verifica quantos elementos tem um group e se for maior que o pode ser manda False se não, no final manda True

riosS :: Terreno -> [(Terreno, [Obstaculo])] -> Bool

Ponto 2, verifica se os rios seguidos tem velocidades opostas, logo sentidos opostos, um da esquerda para a direita e outro da direita para a esquerda

validaPassagemEntreArvores :: (Terreno, [Obstaculo]) -> [(Terreno, [Obstaculo])] -> Bool

Verifica se passa em varias relvas