Tarefa4_2022li1g067

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Função principal

jogoTerminou :: Jogo -> Bool

A função jogoTerminou é a principal deste módulo e responsável por verificar se o jogo terminou e o jogador perdeu.

A função é definida da seguinte forma:

jogoTerminou (Jogo (Jogador (x,y)) (Mapa l ll)) = x || x= l || y || y= length ll || caiuNaAgua (x,y) (ll !! y) || debaixoCarro (x,y) (ll !! y)

O jogo termina se se verificar pelo menos uma das seguintes condições:

o jogador saiu do mapa;

o jogador caiu na água;

o jogador foi atingido por um carro.

Para verificar se o jogador saiu do mapa compara-se as suas coordenadas com as dimensões do mapa na função principal.

As duas últimas condições são verificadas com recurso a 2 funções auxiliares, caiuNaAgua e debaixoCarro.

Funções auxiliares

caiuNaAgua :: Coordenadas -> (Terreno, [Obstaculo]) -> Bool

A função caiuNaAgua retorna False caso o jogador não esteja num rio, ou se está em cima de um tronco ou True se estiver num rio em um sítio sem obstáculos.

caiuNaAgua (x,y) (Rio v, os) | os !! x == Nenhum = True
                             | otherwise = False
caiuNaAgua _ _ = False

debaixoCarro :: Coordenadas -> (Terreno, [Obstaculo]) -> Bool

A função debaixoCarro retorna True se na posição do jogador existir um Carro ou False caso contrário

debaixoCarro (x,y) (Estrada v, os) | os !! x == Carro = True
                                   | otherwise = False
debaixoCarro _ _ = False